pyproject.toml
[project]
name = "web-diagnostics-orchestration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.125.0",
    "google-analytics-data>=0.19.0",
    "httpx>=0.28.1",
    "openai>=2.13.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
